refactor!: move each TS constants & Types into their respective packages
