+++ packages/LSP1UniversalReceiver/index.ts
@@ -1,0 +1,1 @@
+export * from '../constants';
