+++ packages/LSP7DigitalAsset/constants.ts
@@ -1,0 +1,15 @@
+export const INTERFACE_ID_LSP7 = '0xb3c4928f';
+
+export const LSP7_TYPE_IDS = {
+  // keccak256('LSP7Tokens_SenderNotification')
+  LSP7Tokens_SenderNotification:
+    '0x429ac7a06903dbc9c13dfcb3c9d11df8194581fa047c96d7a4171fc7402958ea',
+
+  // keccak256('LSP7Tokens_RecipientNotification')
+  LSP7Tokens_RecipientNotification:
+    '0x20804611b3e2ea21c480dc465142210acf4a2485947541770ec1fb87dee4a55c',
+
+  // keccak256('LSP7Tokens_OperatorNotification')
+  LSP7Tokens_OperatorNotification:
+    '0x386072cc5a58e61263b434c722725f21031cd06e7c552cfaa06db5de8a320dbc',
+};
