+++ packages/lsp17contractextension-contracts/constants.ts
@@ -1,0 +1,7 @@
+export const INTERFACE_ID_LSP17Extendable = '0xa918fa6b';
+export const INTERFACE_ID_LSP17Extension = '0xcee78b40';
+
+export const LSP17DataKeys = {
+  // bytes10(keccak256('LSP17Extension')) + bytes2(0)
+  LSP17ExtensionPrefix: '0xcee78b4094da860110960000',
+};
