+++ packages/lsp1delegate-contracts/constants.ts
@@ -1,0 +1,3 @@
+// ERC165 Interface ID
+// ----------
+export const INTERFACE_ID_LSP1DELEGATE = '0xa245bbda';
