+++ packages/lsp1delegate-contracts/index.ts
@@ -1,0 +1,1 @@
+export * from './constants';
