+++ packages/lsp25-contracts/constants.ts
@@ -1,0 +1,7 @@
+export const INTERFACE_ID_LSP25 = '0x5ac79908';
+
+/**
+ * @dev LSP25 version number for signing `executeRelayCall(...)` transaction using EIP191
+ * @see for details see: https://github.com/lukso-network/LIPs/blob/main/LSPs/LSP-25-ExecuteRelayCall.md#executerelaycall
+ */
+export const LSP25_VERSION = 25;
